+++ NEWS.md
@@ -1,3 +1,3 @@
 # sjosmooth 0.1.0
 
-* Added a `NEWS.md` file to track changes to the package.
+* First release. The package performs kernel smoothing for regresion models with one or more independent variables.  Multiple kernels are supported. Cox proportional hazards regression is the only model officially support/tested (\code{survival::coxph}). Others will be added in subsequent updates.